pyproject.toml
[tool.poetry]
name = "depp"
version = "0.1.0"
description = ""
authors = ["Aadit Rahul Kamat <aadit.k12@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click8 = "^8.0.1"
poetry = "^1.1.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
